src/ai/rl_agent.py
from stable_baselines3.common.vec_env import DummyVecEnv

class RLTradingAgent:
    """Reinforcement learning trading agent."""

    def __init__(self, env):
        """Initialize the agent with the given environment."""
        self.env = DummyVecEnv([lambda: env])

    def get_signal(self):
        """Simulate signal generation by the agent."""
        reset_result = self.env.reset()

        # Updated: Handle the possibility of 2 or 3 values being returned from reset()
        if isinstance(reset_result, tuple):
            if len(reset_result) == 2:
                observation, _ = reset_result
            elif len(reset_result) == 3:
                observation, _, _ = reset_result  # Discard extra values if present
        else:
            observation = reset_result

        return observation  # Placeholder for signal generation

    def train(self):
        """Train the RL agent."""
        pass  # Placeholder for training logic
