tests/test_rl_agent.py
import unittest
from unittest.mock import Mock
from src.ai.rl_agent import RLTradingAgent

class MockEnvironment:
    def reset(self):
        # Simulate different possible return values
        return [0.0, 0.0], {}, {}  # Returning 3 values

class TestRLAgent(unittest.TestCase):

    def setUp(self):
        self.env = MockEnvironment()
        self.agent = RLTradingAgent(env=self.env)

    def test_get_signal(self):
        observation = self.agent.get_signal()
        self.assertIsInstance(observation, list)
        self.assertEqual(observation, [0.0, 0.0])

    def test_train_method(self):
        # Since train is a placeholder, we just test that it can be called
        try:
            self.agent.train()
        except Exception as e:
            self.fail(f"Train method raised an exception: {e}")

if __name__ == '__main__':
    unittest.main()
